v1.4.0: Security and stability overhaul
- Rewrite explosion system to spread block destruction across ticks,
  preventing server freezes on large detonations
- Cap concurrent large explosions at 3, silently drop excess
- Lower max explosion radius from 1000 to 128 (default 64)
- Wire up ENABLE_DNU_VARIANTS config to hide DNU bombs from creative tab
- Fix ClientEvents event bus (MOD bus for renderer/overlay registration)
- Fix all 60 blockstate JSONs to include unstable property variants
- Change NBT fallback from Tsar Bomba to Fritz X (weakest bomb)
- Cache client-side BombData lookup to avoid per-frame string parsing
- Merge FuseOverlay entity iteration into single pass
- Split config into COMMON + CLIENT specs (infographic is per-player)
- Bounds-check fuse on NBT deserialization (clamp 1-600)
- Remove unnecessary short casts in fuse handling
- Harden .gitignore against credential leaks

+++ src/main/java/com/historicbombs/block/HistoricTNTBlock.java
@@ -85,7 +85,7 @@ public class HistoricTNTBlock extends Block {
         HistoricPrimedTNTEntity entity = new HistoricPrimedTNTEntity(level,
             pos.getX() + 0.5, pos.getY(), pos.getZ() + 0.5, explosion.getIndirectSourceEntity(), bombData);
         int fuse = entity.getFuse();
-        entity.setFuse((short)(level.random.nextInt(fuse / 4) + fuse / 8));
+        entity.setFuse(level.random.nextInt(fuse / 4) + fuse / 8);
         level.addFreshEntity(entity);
     }
